scripts/undercurrent/requirements.txt
yfinance>=0.2.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
anyio>=4.0.0
